fix: unitialized snapshot

+++ internal/domains/features/ema.go
@@ -8,23 +8,19 @@ import (
 
 func init() {
 	reg.Features.Register("ema", func(opts map[string]any) (proto.Feature, error) {
-		period, ok := opts["period"].(int)
+		period, ok := opts["period"].(float64)
 		if !ok {
 			return nil, fmt.Errorf("ema register: missing or invalid period")
 		}
 
-		id := iDWithPeriod("ema", period)
+		id := iDWithPeriod("ema", int(period))
 
 		return &EMA{
 			id:     proto.NewKey[float64](id),
-			period: period,
+			period: int(period),
 		}, nil
 	})
 }
-
-// ============================================================================
-// EMA - Exponential Moving Average
-// ============================================================================
 
 type EMA struct {
 	id     proto.Key[float64]
